+++ app/routes/components/Input/changelog.codegen.ts
@@ -19,4 +19,6 @@ export const changelog = ''
   + '- 2025-11-20: Support the `width` prop and let `className` and `style` props\n'
   + '  effect the `InputBase` component rather than the internal `input` element.\n'
   + '- 2025-12-07: Use `IbsBase` component instead of `InputBase` component.\n'
+  + '- 2025-12-18: Use `useOptionalFieldContext` hook to get the field context for\n'
+  + "  the `id` prop's default value. Support `status` prop to set the error state.\n"
   ;