+++ app/routes/components/Input/demos/WithDisabledState/index.tsx
@@ -4,9 +4,8 @@ import { Input } from "@/components/Input";
 export default function Demo() {
   return (
     <Flex dir="column" gap="1rem">
-      <Input size="sm" placeholder="Type something..." />
-      <Input size="md" placeholder="Type something..." />
-      <Input size="lg" placeholder="Type something..." />
+      <Input disabled placeholder="Disabled input" />
+      <Input disabled variant="filled" placeholder="Disabled filled input" />
     </Flex>
   );
 }