+++ src/components/Input/input.tsx
@@ -2,22 +2,27 @@ import classNames from "classnames";
 import type { ComponentProps, FC } from "react";
 
 import { useJss, useTheme } from "@/components/CopyUiProvider";
+import { useOptionalFieldContext } from "@/components/Field";
 import { extractStylesProps, IbsBase } from "@/components/IbsBase";
 
-import styles from "./input.module.scss";
+import styles from "./input.module.css";
 
 export type InputProps = ComponentProps<typeof IbsBase> &
-  ComponentProps<"input"> & {
+  Omit<ComponentProps<"input">, "size"> & {
     variant?: "default" | "filled";
     size?: "xs" | "sm" | "md" | "lg" | "xl";
+    status?: "error";
     leftSection?: React.ReactNode;
     rightSection?: React.ReactNode;
   };
 
 export const Input: FC<InputProps> = (props) => {
   const {
+    id: originalId,
+
     variant = "default",
     size = "md",
+    status,
     leftSection,
     rightSection,
     className,
@@ -30,17 +35,22 @@ export const Input: FC<InputProps> = (props) => {
     ...others
   } = props;
 
+  const fieldContext = useOptionalFieldContext();
+  const id = originalId ?? fieldContext?.id;
+
   const { stx: baseStyleStx, rest } = extractStylesProps(others);
 
   const theme = useTheme();
   const jss = useJss();
 
   const baseStx = jss.hash({
-    "--input-focus-bdColor": theme.colors.blue["800"],
-    "--input-paddingInlineStart": leftSection ? "0.125rem" : "0.75rem",
-    "--input-paddingInlineEnd": rightSection ? "0.125rem" : "0.75rem",
+    "--input-bdColor": status === "error" ? theme.colors.red["650"] : undefined,
+    "--input-pl": leftSection ? "0.125rem" : "0.75rem",
+    "--input-pr": rightSection ? "0.125rem" : "0.75rem",
     "--input-placeholderColor": theme.tokens.inputBasePlaceholderColor,
     "--input-caretColor": theme.colors.blue["800"],
+    "--input-focus-bdColor":
+      status === "error" ? theme.colors.red["650"] : theme.colors.blue["650"],
   });
 
   const inputStx = jss.hash({
@@ -60,6 +70,7 @@ export const Input: FC<InputProps> = (props) => {
       {leftSection || rightSection ? (
         <IbsBase.Wrapper>
           <input
+            id={id}
             className={classNames(styles.input, inputStx)}
             placeholder={placeholder}
             onChange={onChange}
@@ -68,6 +79,7 @@ export const Input: FC<InputProps> = (props) => {
         </IbsBase.Wrapper>
       ) : (
         <input
+          id={id}
           className={classNames(styles.input, inputStx)}
           placeholder={placeholder}
           onChange={onChange}